Add formatting functionality
Add FormatScreen.
The screen tells the GamePad to wipe its' data, which brings it back to the "Turn on the TV and the Wii U console, then switch to the input setting for Wii U on the TV" screen.
TV remote will be erased.
Powering with HOME will be disabled.

Pairing can be done as regular. When pairing, the TV remote will not have a placeholder with "Set up TV remote in system settings" - the button will straight up not work unless reconfigured.